Image upload: add sha to initial packet
There is a new feature in Apache Mynewt which keeps track of unfinished
uploads by storing the hash of the image data. When a client sends the
initial upload packet (offset 0) with the hash that matches an
unfinished upload, the device will respond with the offset which the
previous upload left off at.

Since a SHA256 is a lot of bytes, the ImageManager sends only the first
3 bytes of the SHA which should be random enough to verify that the
images are the same.