Prepare for maven central
This commit modifies the gradle build files for hosting on the maven
central repository. The gradle file which is used to upload the archives to
maven central is taken from https://github.com/JakeWharton/timber.

The group details for the project are located in the root
build.properties file. This file contains, among others:
    * Version: 0.6.3-SNAPSHOT
    * Group: io.runtime.mcumgr

The package specific artifact ID and name are located in the module
gradle.properties file. This file contains:
    * Artifact ID: mcumgr-android
    * Name: McuManager Android
    * Package Type: aar