rename to mcumgr-core; breakout ble to mcumgr-ble
Brief notes:
 * Rename module mcumgr-android-lib -> mcumgr-core.
    - Artifact renamed from mcumgr-android -> mcumgr-core.
 * Breakout ble package from core into mcumgr-ble.
    - mcumgr-ble is built as separate artifact which includes core as an
      api dependency.

This change is useful to allow users to reimplement their own transport
layer or potentially add a new transport all together. Furthermore,
since I plan on adding the CoAP BLE transport at some point to the main
repo, this allows me to simply add a new module, mcumgr-coap-ble, rather
than adding it to mcumgr-android and forceing users to add dependencies
which they will not use.

Finally, I can forsee a future where the core could be broken out into
it's own plain java library for use in non-android applications. This
would simplify that process as it is already built as a separate
artifact.