Add open to transport; call open in dfu manager.
This commit adds support for explicitly opening a connection via the
McuManagerTransport. This is most useful for DFU with transport
implementations which do not connect on demand when send is called.

McuManager Core Changes
-----------------------
McuMgrTransport has a new method, void open(); which should be
implemented to open the connection to the remote device if not already.

McuManager BLE Transport Changes
--------------------------------
McuMgrBleTransport has implemeneted the open() method to asynchronously
connect to the device with 3 reties delayed by 100 ms. On success,
connection observers are notified.

Sample App Changes
------------------
The sample app has been changed to connect to the BLE device in
MainActivity onCreate().